AGORA SIM, Versão final CRUD

+++ read.php
@@ -30,7 +30,7 @@
                     <th> CATEGORIA </th>
                     <th>
                         <a href='update.php?id={$row['id_notas']}'>Editar</a> |
-                        <a href='delete.php?id={$row['id_notas']}'>Excluir</a>
+                        <a href='delete.php?id={$row['id_notas']}' onclick='return confirm(\"Tem certeza que deseja excluir esta nota?\");'>Excluir</a>
                     </th>
                 </tr>";
 
@@ -38,10 +38,7 @@
                     
                     <td> {$row['titulo']} </td>
                     <td> {$row['categoria']} </td>
-                    <td>
-                        <a href='update.php?id={$row['id_notas']}'>Editar</a> |
-                        <a href='delete.php?id={$row['id_notas']}' onclick='return confirm(\"Tem certeza que deseja excluir esta nota?\");'>Excluir</a>
-                    </td>
+                    
                 </tr>
                 <tr>    
                     <th> CONTEUDO </th>
